refactor: Rename crate
