+++ Cargo.toml
@@ -44,7 +44,7 @@ members = [
     "crates/copilot_chat",
     "crates/crashes",
     "crates/credentials_provider",
-    "crates/csv_preview",
+    "crates/tabular_data_preview",
     "crates/dap",
     "crates/dap_adapters",
     "crates/db",
@@ -318,7 +318,7 @@ copilot_ui = { path = "crates/copilot_ui" }
 crashes = { path = "crates/crashes" }
 credentials_provider = { path = "crates/credentials_provider" }
 crossbeam = "0.8.4"
-csv_preview = { path = "crates/csv_preview"}
+tabular_data_preview = { path = "crates/tabular_data_preview"}
 dap = { path = "crates/dap" }
 dap_adapters = { path = "crates/dap_adapters" }
 db = { path = "crates/db" }
