+++ crates/tabular_data_preview/LICENSE-GPL
