+++ crates/tabular_data_preview/src/renderer/performance_metrics_overlay.rs
