+++ crates/tabular_data_preview/src/renderer/preview_view.rs
