+++ crates/tabular_data_preview/src/renderer/render_table.rs
