+++ crates/tabular_data_preview/src/renderer/row_identifiers.rs
