+++ crates/tabular_data_preview/src/renderer/settings.rs
