+++ crates/tabular_data_preview/src/renderer/table_cell.rs
