+++ crates/tabular_data_preview/src/renderer/table_header.rs
