+++ crates/tabular_data_preview/src/settings.rs
