+++ crates/tabular_data_preview/src/table_data_engine.rs
