+++ crates/tabular_data_preview/src/tabular_data_preview.rs
