+++ crates/tabular_data_preview/src/types.rs
