+++ crates/tabular_data_preview/src/types/coordinates.rs
