+++ crates/tabular_data_preview/src/types/table_cell.rs
