+++ crates/tabular_data_preview/src/types/table_like_content.rs
