+++ crates/zed/Cargo.toml
@@ -93,7 +93,7 @@ copilot.workspace = true
 copilot_chat.workspace = true
 copilot_ui.workspace = true
 crashes.workspace = true
-csv_preview.workspace = true
+tabular_data_preview.workspace = true
 dap_adapters.workspace = true
 db.workspace = true
 debug_adapter_extension.workspace = true
